Refactor param schedulers and make them generic (#324)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/324

Made classy param schedulers generic so that they can be used to schedule any parameter and not just "lr".

The optimizer config can now have an optional "param_schedulers" key which would contain a dict of param scheduler config for various params.
A normal config would change from -
```
"optimizer": {
    "lr": {
         ...
    }
    ...
}
```
to -
```
"optimizer": {
    "param_schedulers": {
        "lr": {
             ...
        }
    }
    ...
}
```
To schedule lr and momentum, the config would look like this -
```
"optimizer": {
    "param_schedulers": {
        "lr": {
             ...
        },
        "momentum": {
             ...
        }
    }
    ...
}
```

The optimizers required an lr_scheduler, this requirement has been removed. `ClassyOptimizer`s can be initialized just like `PyTorch` optimizers.

Added a `set_param_schedulers()` function to `ClassyOptimizer`. This helps with the next improvement.

All the optimizers had to build the schedulers inside their `from_config` methods because of the way the old setup worked. Now, the `build_optimizer()` function also builds the param schedulers. This is done by building the optimizer, param schedulers and then attaching the param schedulers to the optimizer using `set_param_schedulers()`

Added a default values for all the parameters in `__init__()` and in `from_config()`.

We still support the old format of specifying the lr_scheduler, have added a warning that it is deprecated and will be removed in v 0.2.

Reviewed By: vreis

Differential Revision: D18946367

fbshipit-source-id: a5521f878d118c6f652deb41423745cb07cf2962