do not overwrite multiprocessing context to be fork in ClassyVideoDataset (#328)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/328

- We were forced to use `fork` mp context because `VideoClips` class instance uses too many `torch.Tensor` internally, and each `torch.Tensor` requires to use one File Descriptor (FD) when `VideoClips` instance is shared by main process and worker process in the PyTorch dataloader.
OS has a max limit of FD (by default 65K), which is often exceeded in practice.

We fix this issue in `VideoClips` in previous diff D19173248. Therefore, we no longer need to overwrite the mp context to be `fork` in `ClassyVideoDataset`. This avoid duplicating dataset instances, and reduce the CPU memory footprint.

We won't have the hanging issues caused by the `fork` mp context which is not recommended to use at Facebook.

Differential Revision: D19173397

fbshipit-source-id: 0b89f8df33c51ccf4f497a9392170e6a7ab7358e