Updated GenericTransform to have more consistent behavior for None transforms
Summary:
GenericTransform had inconsistent behavior when transform = None. For most transforms, this would mean "Do not transform" the image, for GenericTransform it would throw an error.

This diff fixes this by allowing both the split AND the transform argument to be None. In that case, the torchvision tuple -> dict change is still done.

Partially addresses https://github.com/facebookresearch/ClassyVision/issues/334

Differential Revision: D19522477

fbshipit-source-id: 208081b257e645ac8ed08d5feb347af42bc9fe3c