Hydra improvements (#58)
Summary:
Pull Request resolved: https://github.com/fairinternal/ClassyVision/pull/58

The original hydra `args.yaml` was part of the classy vision package, which meant editing it in a classy project wasn't possible without changing the code.

This diff makes changes so that the `args.yaml` is generated inside the project, giving users full control to the config.

These are the changes being made -
- Removed `classy_vision.generic.args` - the logic has been moved to `classy_train.py`. This was needed because otherwise the search happens inside  the package and not in the file system. Also, if we only have hydra then there is almost no additional code.
- Moved configs inside `classy_vision/hydra` to `classy_vision/hydra/conf`
- Moved `args.yaml` to `templates/synthetic/hydra_configs`
- Added a `hydra_plugins` package which tells hydra to add these bundled configs to the search path

See the test plan for usage examples.

NOTE: These changes are made with hydra-core 0.11 (latest release) in mind and will not work with the hydra master code.

Reviewed By: vreis

Differential Revision: D19418196

fbshipit-source-id: c2a5f046de75dde1eaf76ee0166308e30d47a9ac