Updated GenericTransform to have more consistent behavior for None transforms (#59)
Summary:
Pull Request resolved: https://github.com/fairinternal/ClassyVision/pull/59

Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/355

GenericTransform had inconsistent behavior when transform = None. For most transforms, this would mean "Do not transform" the image, for GenericTransform it would throw an error.

This diff fixes this by allowing both the split AND the transform argument to be None. In that case, the torchvision tuple -> dict change is still done.

Partially addresses https://github.com/facebookresearch/ClassyVision/issues/334

Reviewed By: mannatsingh

Differential Revision: D19522477

fbshipit-source-id: 73b24aa1227bcedcfe206438c53e15f730edd664