Create a loss that deals with multiple outputs by using the first output only
Summary:
We need FAIM model to output attention weights as well, so we can use them for localizations. For instance, we can save the to hive with WriteToHiveHook and evaluate.

However classy vision passes model output directly to loss function, which in most cases assumes a single output:  https://fburl.com/diffusion/68zqh9m6

To get around this, we need a loss that just takes the first output from the model.

Differential Revision: D19582685

fbshipit-source-id: 53d28cea7dbf5e77d2284896e557826a2fc4a3bf