Add RMSPropTF optimizer
Summary:
Added a version of rmsprop, `RMSPropTF`, which uses the epsilon inside the square root, like in tensorflow. This helps reproducing papers implemented in tf.

I copied the implementation and didn't fix the lint error so that it's easier to find the differences between the original and this class, `RMSpropTFOptimizer`.

Reviewed By: vreis

Differential Revision: D19667085

fbshipit-source-id: 4181fa88f581ad81937dc2c87682e4534e7e17fc