Checkpoint only at the end of train phases, only run on_start on master
Summary:
We checkpoint after test phases by default, which doubles the memory consumption for checkpoints and it doesn't feel worth it to get the ability to resume after the test phases. Changed the default to train only.

During `on_start` we check if the checkpoint folder is present, made it so that the check only happens on the master worker.

Differential Revision: D19770656

fbshipit-source-id: 6522be85154db50dc3dfbc0d92fb2ed72c3a5f7c