Checkpoint only at the end of train phases, only run on_start on master (#378)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/378

We checkpoint after test phases by default, which doubles the memory consumption for checkpoints and it doesn't feel worth it to get the ability to resume after the test phases. Changed the default to train only.

During `on_start` we check if the checkpoint folder is present, made it so that the check only happens on the master worker.

Reviewed By: vreis

Differential Revision: D19770656

fbshipit-source-id: 4a53e528fd944db123a1f254530fe7a036b413a5