Take a separate argument for loading checkpoints, take paths for pretrained checkpoints (#379)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/379

There are now two arguments for checkpoints -
- `checkpoint_folder`: This is where checkpoints are saved to. Checkpoints are saved inside a subdirectory (the flow id) within this folder
- `checkpoint_load_path`: This is where checkpoints are loaded from. This can be a file or a directory.

Note that this means that training will not automatically resume by picking up a checkpoint from the `checkpoint_folder` if the user doesn't specify a `checkpoint_load_path`, but I think explicit is better than implicit in this scenario.

Also, renamed `pretrained_checkpoint_folder` to `pretrained_checkpoint_path` and updated the help string to reflect the fact that this doesn't need to be a folder anymore.

Updated the Getting started tutorial to reflect the changes.

Reviewed By: vreis

Differential Revision: D19760255

fbshipit-source-id: 7c66c7b66c0d4a192dcc6b403852668d01f7932b