Add support to attach heads to DenseNets (#383)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/383

Added the following attachable blocks to the `DenseNet` implementation-
Contains the following attachable blocks -
- `block{block_idx}-{idx}`
- `transition-{idx}`
- `trunk_output`

The `trunk_output` block is the final output of the `DenseNet`. This is where a `fully_connected` head will normally be attached.

Also added an additional arg (`zero_init_bias`) to `FullyConnectedHead` since `DenseNets` init the fc bias to 0.

Reviewed By: vreis

Differential Revision: D19780372

fbshipit-source-id: f52bafc7a6d03f4f6678d2726918b23769ddeb3f