extend profiler to allow module customized method of flops computing
Summary:
Current profiler uses condition branch based on layer type to define the way of flops computing. As a centralized place containing all implementation, It may not be scalable as the number of layer types grows.

Thus, extend it to allow module customized method of flops computing defined in the original module.

Differential Revision: D19819686

fbshipit-source-id: 1ad28d11da2238a617ecf94035bcc9284d4deebd