extend profiler to allow module customized method of flops computing (#389)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/389

Current profiler uses condition branch based on layer type to define the way of flops computing. As a centralized place containing all implementation, It may not be scalable as the number of layer types grows.

Thus, extend it to allow module customized method of flops computing defined in the original module.

Reviewed By: mannatsingh

Differential Revision: D19819686

fbshipit-source-id: 49c603bc3231c828e864c24c15b557b2cf3203f7