Support losses with learned parameters
Summary:
This adds support for stateful losses with learned parameters. Examples include the NormFace / SphereFace / CosFace / ArcFace family of losses, or noise-contrastive estimation losses that learn partition parameters.

Loss parameters are added to the optimizer, and saved to checkpoints. Losses with parameters are wrapped in a separate DistributedDataParallel instance to synchronize parameter initialization, and gradients, across nodes.

Differential Revision: D19717056

fbshipit-source-id: bc4eec5e4d3af6072943b963ad4f8d904e7c936c