Support losses with learned parameters (#394)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/394

This adds support for stateful losses with learned parameters. Examples include the NormFace / SphereFace / CosFace / ArcFace family of losses, or noise-contrastive estimation losses that learn partition parameters.

Loss parameters are added to the optimizer, and saved to checkpoints. Losses with parameters are wrapped in a separate DistributedDataParallel instance to synchronize parameter initialization, and gradients, across nodes.

Reviewed By: vreis

Differential Revision: D19717056

fbshipit-source-id: 64b815b1890ee8029517f5ed1a9edf8cff2ba0e7