Remote on_update hook call from train_step
Summary:
We want people to write their own train_step functions, and having a hook in
there makes that complicated. on_update was already happening at the end of
train_step, so move it out into the step function.

Another subtle change here is that we now call on_step for both train and test
phases.

Differential Revision: D19906465

fbshipit-source-id: d5a8702efbc54f2efb8e39802e4df2be2695b0d7