Remote on_update hook call from train_step (#399)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/399

We want people to write their own train_step functions, and having a hook in
there makes that complicated. on_update was already happening at the end of
train_step, so move it out into the step function.

Another subtle change here is that we now call on_step for both train and test
phases.

Reviewed By: aadcock

Differential Revision: D19906465

fbshipit-source-id: 9e4a094e1aa2f9b2c2ddd4a72554febe05b715e4