Support LARC for SGD optimizer only in classy vision (#408)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/408

Pull Request resolved: https://github.com/fairinternal/ClassyVision/pull/64

In an attempt to implement SimpleCLR for contrastive losses, I needed LARC to enable large batch training. mannatsingh had already done work on this during classy vision open source release. https://our.intern.facebook.com/intern/diff/D18542126/

I initially tried using that diff to have a separate standalone LARC to work for any optimizer but it turned out to be tricky to setup correctly as we need to wrap a given optimizer in LARC (the `getattr` and `setattr` functions were not working). I talked to vreis about it and we decided that for now, we can support it for SGD only, file a task to support other optimizers too later after discussions with mannatsingh once he's back.

Reviewed By: vreis

Differential Revision: D20139718

fbshipit-source-id: 7e9124ac927e62b11bd39e3e8ac612ed06f16c20