Add hooks building to classification task (#402)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/402

Pull Request resolved: https://github.com/fairinternal/ClassyVision/pull/62

Add configurable hooks to classification task, had to remove the typehints from ClassyHook to avoid a circular dependency. As discussed, there are other options (for example importing in from_config), but since we keep running into this issue we are going to remove them to try and improve dev velocity.

One more note, as requested I made the checkpoint hook a default option, but this led to some additional changes (need to allow input_args in checkpoint hook to be None as well as empty dict and had to add a make dir call in the task build).

Reviewed By: vreis

Differential Revision: D19770583

fbshipit-source-id: 758d4693ba882c6c4899ba548a8e2615a4849b6f