Move use_gpu from ClassyTrainer to ClassificationTask (#468)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/468

This is the first in a series of diffs to eliminate the ClassyTrainer abstraction. The only reason Trainer existed was to support elastic training, but PET v0.2 does not require changing out training loop. The plan is to move all attributes from ClassyTrainer into ClassificationTask.

Start by moving use_gpu to the task.

Reviewed By: mannatsingh

Differential Revision: D20801017

fbshipit-source-id: d56bb330737a98a7ea7545d33da0ad0f21a0b6a1