Move num_dataloader_workers out of ClassyTrainer
Summary:
In order to get rid of ClassyTrainer, we need to move num_dataloader_workers and dataloader_mp_context somewhere else. Move the context to the task and num_workers to ClassyDataset. This also changes the default multiprocessing context to spawn, for consistency. That required changing tests in fblearner to dev-nosan.

This is a breaking change to the configs, and I'll announce it in the group before landing. Will also try to land other breaking changes at the same time to minimize disruption.

Differential Revision: D20803692

fbshipit-source-id: 0612e24857c538ff7e4495c2d8efecdc1c5a648b