Revamp logging (#478)
Summary:
This is a bunch of changes to make our training logs more meaningful and
easier to understand. We print the task config in the beginning of training,
make it clear what values are approximate or final, supress verbose logs
by default and format floats accordingly.

Before this diff: P128995674
After this diff: P128995244
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/478

Reviewed By: mannatsingh

Differential Revision: D21022171

Pulled By: vreis

fbshipit-source-id: d63d5ac9b4b3b3cc9abb359914141bf6821dc14a