Move num_dataloader_workers out of ClassyTrainer (#10)
Summary:
Pull Request resolved: https://github.com/fairinternal/ssl_scaling/pull/10

Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/477

In order to get rid of ClassyTrainer, we need to move num_dataloader_workers and dataloader_mp_context somewhere else. Move the context to the task and num_workers to ClassyDataset. This also changes the default multiprocessing context to spawn, for consistency. That required changing tests in fblearner to dev-nosan.

This is a breaking change to the configs, and I'll announce it in the group before landing. Will also try to land other breaking changes at the same time to minimize disruption.

Reviewed By: mannatsingh

Differential Revision: D20803692

fbshipit-source-id: f5091bd93b3265271411f6dcd8d995b56c1cd565