Fix profiler infinite recursion bug for re-used modules
Summary:
Found another bug in the profiler - this one happens while patching the modules for FLOPs / activations calculations. The original logic used to recursively iterate over the children of a module. The problem with that is that if a module is re-used (see added test case for example), it'll be the child of more than one modules. This would mean that we patch it more than once, causing the infinite recursion bug.

The fix is to iterate over all the modules using `model.named_modules()`, which only lists a module once.

Added a test case which failed before the fix.

Differential Revision: D21168917

fbshipit-source-id: 90d7d34ab42a94eb69b69c56d851acf9611c3fbf