Fix profiler infinite recursion bug for re-used modules (#487)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/487

Found another bug in the profiler - this one happens while patching the modules for FLOPs / activations calculations. The original logic used to recursively iterate over the children of a module. The problem with that is that if a module is re-used (see added test case for example), it'll be the child of more than one modules. This would mean that we patch it more than once, causing the infinite recursion bug.

The fix is to iterate over all the modules using `model.named_modules()`, which only lists a module once.

Added a test case which failed before the fix.

Reviewed By: aadcock

Differential Revision: D21168917

fbshipit-source-id: b08276336f2029323b3fd50b44e2aff542c0c51d