allow 'num_samples` to be None in OnBoxVideoClipsDataset
Summary:
Currently in `OnBoxVideoClipsDataset`, we need to set `num_samples` ahead of time, and at running time `DataloaderLimitWrapper` will try to return so many samples, potentially with duplicate ones.

This is not suitable in some use cases. For example, for feature extraction via `write_to_hive_hook`, we may not know the no. of samples ahead of time. So the actual no. of samples depends on other hyper-parameter. For example, if we want to extract clips feature for clips between the second 30 and second 40, and the actual no. of qualified videos depend on the video duration. Say the batch size per model replica is 1. In such case, those videos shorter than 30 seconds will not form minibatch by being skipped by `DataloaderSkipNoneWrapper`. `DataloaderLimitWrapper` will find duplicate qualified videos by wrapping around until meeting `num_samples` videos.

Therefore, we relax such hard requirement, and allow `num_samples` to be `None`, which will disable `DataloaderLimitWrapper` effectively.

This change is back-compatible, and existing users are not expected to be interrupted.

Differential Revision: D21320406

fbshipit-source-id: daa2f6e43607cbdfa27a1910cbefeeac34b5e82f