add verbose argument to ComplexityComputer class
Summary:
Currently, the profiler only reports the overall flops and activations of a model.
In some cases, such as designing more efficient models, we further need to know per-module flops and activations. Therefore, add an argument `verbose` and when it is true, the profiler also report flops and activations in each module.

The change is back-compatible as the default value of `verbose` is False.

Differential Revision: D21634970

fbshipit-source-id: abb746fe003bc4070c20638c4d4d035efa9b6a8b