add verbose argument to ComplexityComputer class]
Summary:
Currently, the profiler only reports the overall flops and activations of a model.
In some cases, such as designing more efficient models, we further need to know per-module flops and activations. Therefore, add an argument `verbose` and when it is true, the profiler also report flops and activations in each module.

The change is back-compatible as the default value of `verbose` is False, and no new messages will be logged by default in `ModelComplexityHook`.

Differential Revision: D21635217

fbshipit-source-id: 62f137d22de933b7543935261dccb334df003b5f