Implement a Classy hook to convert trained model to a TorchScript (#513)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/513

When we deploy PyTorch model to inference platform, we need to convert the trained model into `TorchScript`.

In this diff, we want to implement a ClassyVision hook, which is triggered by the end of FBLearner workflow and serialize the trained model from checkpoint into TorchScript.

Introduce a new workflow param `torchscript_folder` for user to specify a path to save the files.

Reviewed By: mannatsingh, kazhang

Differential Revision: D21506773

fbshipit-source-id: 858491559cb6d8d2246aee744e2451804f47a16a