add verbose argument to ComplexityComputer class] (#519)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/519

Currently, the profiler only reports the overall flops and activations of a model.
In some cases, such as designing more efficient models, we further need to know per-module flops and activations. Therefore, add an argument `verbose` and when it is true, the profiler also report flops and activations in each module.

The change is back-compatible as the default value of `verbose` is False, and no new messages will be logged by default in `ModelComplexityHook`.

Reviewed By: vreis

Differential Revision: D21635217

fbshipit-source-id: 8373c37f8645d1b79e67ef6d3272be61e8cb85b3