Support sync batch norm group size (#534)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/534

Sync BatchNorm is an expensive operation when we try to synchronize across multiple nodes. Adding the option of specifying the group size for the sync op. A value of 8 would mean that the sync only happens intra node (since the ranks are aligned in multiples of 8) for an 8 GPU per node setup.

This only works with Apex Sync BN. When trying it with PyTorch with a group size of 8, I was getting connection reset errors inside `init_distributed_data_parallel_model`. When the group size equalled the total number of GPUs, I didn't get issues.

I think we have too many options at the config level for distributed settings. Moving them to a separate section in a follow up diff.

Reviewed By: vreis

Differential Revision: D21868629

fbshipit-source-id: e659d8339a2d03a6cb4be2eb9599223583191c2b