parse the input shape recursively in profiler
Summary:
When we print out the input tensor size around L334, we use `x.size()`.
However, there are cases the input `x` is a list/tuple/dict rather than a torch.Tensor, as we have considered in existing function `_get_batchsize_per_replica(...)`.

So fix this issue in the profiler.

Reviewed By: mannatsingh

Differential Revision: D21702925

fbshipit-source-id: 782ac54a5fdda0e743900e91e635ee67039cdb82