Log sample fetch times to tensorboard (#560)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/560

Log cumulative sample fetch times to get an idea about how much time we're spending waiting on the dataloaders.
- The sample fetch time for a given step is calculated as the max of the sample fetch time across all workers
- Log the cumulative sample fetch time per phase type vs the number of batches
- This change now requires the tb hook to be plugged in to all replicas since there is a sync involved

NOTE: I changed the x axis for the LR curve to be batches instead of samples to be consistent with the sample fetch times plot. This should anyway be easier to understand than samples.

Reviewed By: vreis

Differential Revision: D22298279

fbshipit-source-id: 003517b73ce841f20c060072555fe94dfafee082