Make hook checkpoints backwards compatible
Summary:
If we ever add any new state to a hook, the old checkpoints become unusable since the hook crashes. This is because we replace the state's internal `__dict__` and the newly added attributes are erased.

Instead, this diff converts the `set_classy_state` to an `update` which will ensure that such hooks will continue running. There are no guarantees on correctness though.

The other option is to not set the state of a hook at all if there is a key mismatch. We can revisit this if the latter turns out to be a better option.

Differential Revision: D22604083

fbshipit-source-id: 7614068e530a3bb364c57f9ad784d59d4dcad861