Fix ClassyLoss state
Summary:
If the loss has some parameters, then we update self.loss to point to a DistributedDataParallel object.
When the checkpoint is being saved, the get_classy_state function is triggered, which checks if the self.loss is a ClassyLoss instance. This would be false since it is a DDP instance.
The diff fixes this behavior.

Reviewed By: kazhang

Differential Revision: D22634936

fbshipit-source-id: c79cf15a1497d6fab50ee4da7148284f7218271a