Fix ClassyLoss state (#50)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/50

Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/579

If the loss has some parameters, then we update self.loss to point to a DistributedDataParallel object.
When the checkpoint is being saved, the get_classy_state function is triggered, which checks if the self.loss is a ClassyLoss instance. This would be false since it is a DDP instance.
The diff fixes this behavior.

Reviewed By: kazhang, prigoyal

Differential Revision: D22634936

fbshipit-source-id: 6d827365c136809a6bced3a29b0749d48038d258