Make hook checkpoints backwards compatible (#567)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/567

If we ever add any new state to a hook, the old checkpoints become unusable since the hook crashes. This is because we replace the state's internal `__dict__` and the newly added attributes are erased.

Instead, this diff converts the `set_classy_state` to an `update` which will ensure that such hooks will continue running. There are no guarantees on correctness though.

The other option is to not set the state of a hook at all if there is a key mismatch. We can revisit this if the latter turns out to be a better option.

Reviewed By: kazhang

Differential Revision: D22604083

fbshipit-source-id: 9f95000ae2009329d068d45cd2b4816e4ccb5207