Reset wrapper cls in a finally block (#587)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/587

We reset the `wrapper_cls` every time we modify it, but that only works if the tests pass. If any such test fails, the state is modified which crashes other tests too.

All this diff does is wrap the `wrapper_cls` reset logic in a finally block so that the state is fixed even if the tests fail.

Reviewed By: vreis

Differential Revision: D22908275

fbshipit-source-id: edda023713f645df8fa3211537530b340c711bb4