reduce memory use of PreciseBatchNormHook
Summary:
In current `PreciseBatchNormHook`, it caches samples which are used to compute precise BN at the end of phase. The memory use will increase significantly.

For example, in video classification, when cache size is 10,000, input frame is 16, spatial resolution is 224, and no. of gpus is 8, the total cache size is (10000 * 16 * 3* 224 * 224 * 8) = 192 GB on a single host, which is hard to afford.

Thus, improve the implementation to remove the use of cache.

Differential Revision: D23191690

fbshipit-source-id: a4443e0544b3b4ff7033230e7914ea0fccfa4a8b