Move optimizer.on_epoch call to task.on_phase_end (#604)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/604

According to `optimizer.on_epoch()`, it should be called at the end of a phase. Instead, it was being called in `task.advance_phase()`. Moved the to a more appropriate location inside `task.on_phase_end()`

Since `on_epoch` is now called after the final phase as well, it calls `self._update_schedule()` only if `where < 1`. This is guaranteed to not suppress any bugs since the `step()` call still checks for `where < 1`.

Moved the `where < 1` check to the `step()` call directly instead of the indirect call via `_update_schedule()`. This doesn't impact anything but it's easier to reason about - you cannot perform an optimization step if training is finished.

Differential Revision: D23190096

fbshipit-source-id: 9ef80dafa594ec5c57125a4ec31853baa3aa157c