allow to load hook state dict in FineTuningTask
Summary:
Currently, in `FineTuningTask`, we only load the state dict of the base model from pre-trained checkpoint.

This diff allows to selectively load state dict of hooks.

This is useful in some cases. For example, when fine-tuning a model with newly inserted blocks (e.g. GloRe), we want to restore the EMA hook from the pre-trained checkpoint.

Differential Revision: D24880780

fbshipit-source-id: b4e78d14c583853c2913253d7836f83505b911be