allow to load hook state dict in FineTuningTask (#648)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/648

Currently, in `FineTuningTask`, we only load the state dict of the base model from pre-trained checkpoint.

This diff allows to selectively load state dict of hooks.

This is useful in some cases. For example, when fine-tuning a model with newly inserted blocks (e.g. GloRe), we want to restore the EMA hook from the pre-trained checkpoint.

Reviewed By: kazhang

Differential Revision: D24880780

fbshipit-source-id: 6c5ba3043d9bf2915c8f1eee09a810df810f0fa2