reduce memory use of PreciseBatchNormHook (#605)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/605

In current `PreciseBatchNormHook`, it caches samples which are used to compute precise BN at the end of phase. The memory use will increase significantly.

For example, in video classification, when cache size is 10,000, input frame is 16, spatial resolution is 224, and no. of gpus is 8, the total cache size is (10000 * 16 * 3* 224 * 224 * 8) = 192 GB on a single host, which is hard to afford.

Thus, improve the implementation to by adding an argument `cache_sample`.  If false, we do not cache samples, but re-create data loader at the end of training phase.

Reviewed By: mannatsingh

Differential Revision: D23191690

fbshipit-source-id: 3874ef3afcd71b7124a8b047c4be55d6eaab634f