Save CPU memory during broadcast (#616)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/616

Originally, a broadcast caused secondary workers to allocate 3x the memory and the primary to allocated 4x the memory of a checkpoint.
This fix reduces the memory usage to only the memory required to hold the checkpoint on secondary workers (1x), and 3x the memory on the primary.

This can be disabled optionally by passing `use_disk=False` to `broadcast_object()`.

Differential Revision: D23414229

fbshipit-source-id: 4d3b1ce128ef581fb71077d8436e2b563b6de9b5