Progressbar hook with tqdm
Summary:
Moves the progressbar hook to tqdm.

The new hook lets tqdm estimate the remaining time for the overall task, and for the current phase in a given epoch (train or test). Processing times are displayed in terms of images/second as a postfix of the progress bars.

The new progress bar assumes tasks more similar to `ClassificationTask`. In particular it accesses the `task.last_batch.output` field to obtain the size of the batch in number of images.

Differential Revision: D25363238

fbshipit-source-id: 9587215f81c4c8e40051068d133d0a92caaac512