Implement config validation to find unused keys (#665)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/665

Implement a `ClassyConfigDict` type which supports tracking reads and freezing the map (the latter is unused currently).

Added it to `build_task` to catch cases where we don't use any keys passed by users.

This will not catch all instances, like when some components do a deepcopy - we assume all the keys and sub-keys are read in such a situation

Differential Revision: D25321360

fbshipit-source-id: c98b99977cfcbf98c669c0da29b10320ec7efc77