Fix broken Sharded SGD tests (#724)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/724

The sharded SGD tests have been broken for a while because the ports conflict during distributed init. Moved to file based init.

Changed the dep installation order in circle ci because fairscale (a dev dep) was bumping torch to a higher version. Instead, we should just fail in such a scenario.

Differential Revision: D27235293

fbshipit-source-id: e9c2d056334feaa98860a9e6b77840d038ec7d68