Fix broken Sharded SGD tests (#724)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/724

The sharded SGD tests have been broken for a while because the ports conflict during distributed init. Moved to file based init.

Changed the dep installation order in circle ci because fairscale (a dev dep) was bumping torch to a higher version. Instead, we should just fail in such a scenario.

Reviewed By: kazhang

Differential Revision: D27235293

fbshipit-source-id: 95baab53f658b922473ca7c7c02a55ef5ab835ae