ClassyVision Changes to Support checkpoint_load_strict Flags (#98)
Summary:
Pull Request resolved: https://github.com/fairinternal/ClassyVision/pull/98

Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/757

It is very hard to extract features from the intermediate layers of a classy model because `checkpoint_load_strict` is hardcoded to True. This diff allows for setting that to false when user only constructs part of the model while loading the full model state dict to get intermediate layer features.

The flag is default to True so it will not affect existing workflows.

Reviewed By: sallysyw

Differential Revision: D29948213

fbshipit-source-id: 07b585bd75c77521d8651de47c3168fbd84ad7c3