docs: add social banner in support of Ukraine (#782)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

<img width="1080" alt="image" src="https://user-images.githubusercontent.com/12485205/156686642-f8171f22-8aba-4af0-b4dc-25e798f65bb0.png">

Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/782

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

**Static Docs Preview: classyvision**
|[Full Site](https://our.intern.facebook.com/intern/staticdocs/eph/D34633048/V2/classyvision/)|

|**Modified Pages**|

Reviewed By: mannatsingh

Differential Revision: D34633048

Pulled By: dmitryvinn-fb

fbshipit-source-id: 7388fd1f287682b9bfbf92169975d36f85e07f21