Add stateless output option for classy block (#101)
Summary:
X-link: https://github.com/fairinternal/ClassyVision/pull/101

This diff adds a stateless output option within `ClassyBlock.py` when if the environmental variable `STATELESS_CLASSY_OUTPUT` is enabled, `self.output` within `ClassyBlock` never hold the value of the last output (and thus is stateless).  This is useful to prevent certain race conditions, while the default condition will still be maintained if the flag is not enabled to maintain original behavior.

Differential Revision: D35662345

fbshipit-source-id: 8b007b656b68ee23ca3e755b0310406305b02d77