Add stateless output option for classy block (#785)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/785

X-link: https://github.com/fairinternal/ClassyVision/pull/101

This diff adds a stateless output option within `ClassyBlock.py` when if the environmental variable `CLASSY_BLOCK_STATELESS` is set to "1", `self.output` within `ClassyBlock` never hold the value of the last output (and thus is stateless).  This is useful to prevent certain race conditions, while the default condition will still be maintained if the flag is not enabled to maintain original behavior.

Reviewed By: mannatsingh

Differential Revision: D35662345

fbshipit-source-id: af2a4669c96f1c01454602ebd63de695632adc8a