Fix classy_vision test on heads_fully_connected_head_test
Summary:
The heads_fully_connected_head_test on classy_vision seems to failed, see: https://www.internalfb.com/intern/test/562949985032390/

And this can be reproduced in on-demand server by running the buck command provided on the above url.

The error message:
```
RuntimeError: linalg.norm: If dim is specified, it must be of length 1 or 2. Got [1, 2, 3]
```
indicate that the test has wrong value of `dim` input.

This diffs try to solve this issue

Differential Revision: D36702425

fbshipit-source-id: 1a6afacf0584f9bd06d4a587d14ed6fa18c28a12