Fix classy_vision test on heads_fully_connected_head_test (#792)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/792

The heads_fully_connected_head_test on classy_vision seems to failed, see: https://www.internalfb.com/intern/test/562949985032390/

And this can be reproduced in on-demand server by running the buck command provided on the above url.

The error message:
```
RuntimeError: linalg.norm: If dim is specified, it must be of length 1 or 2. Got [1, 2, 3]
```
indicate that the test has wrong value of `dim` input.

This diffs try to solve this issue

Reviewed By: jdsgomes

Differential Revision: D36702425

fbshipit-source-id: ec473f58891bcf6767bb3d8b74b4bf16b904de59