Implement resize and train XRayVideo A/V with only resizing (#796)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/796

We want to check whether training XRayVideo with simply video resizing (in addition to other existing transformation like horizontal flipping and normalization) without random corp is sufficient.

The resize dimension is used as 224*224.

workflow: f362077622 (Note: in the workflow `fcc_mvit_dataset_v4p2_arkc.yaml` is used which I renamed to `fcc_mvit_dataset_v4p2_onlyresize.yaml` in this diff.)

As can be seen, the validation MAP goes to around .422 as opposed to 0.46 when random resized crop is used (f355567669) and rest of the configuration is kept the same. Hence, it is better to keep random resized crop.

Differential Revision: D38522980

fbshipit-source-id: 037690a4dccf9c3ee66b353792cb30b22bae8161